Value tooltip when clicked
Description:

Which is?
It is a tooltip that opens when clicked on some chart value. It then displays this value to the user.

Why is it important?
I believe that in some cases it is possible that the user has many values ​​in the graph and at this moment wants to know the exact value of each dot.

Does the component work and is it perfect?
The component is working properly as it should, however I believe you can improve it, especially in its design.