line chart shadow option
Added one option to dataset
shadowColor: similar to color, if not provided, shadow color for the dataset will be the same as the line color with useShadowColorFromDataset:true,
if useShadowColorFromDataset:false,
shadow color will be chartConfig.color.

shadow color is applied with following priority.
dataset.shadowColor -> dataset.color -> chartConfig.color